Fix build issues

+++ Bomberman/LevelComp.cpp
@@ -9,7 +9,7 @@ bm::LevelComp::LevelComp(dae::GameObject& parent)
 	:Component(parent)
 {
 	//create a "2D" grid of tiles and set the pointers of the array to them
-	for (int i{ 0 }; i < m_Tiles.size(); ++i)
+	for (size_t i{ 0 }; i < m_Tiles.size(); ++i)
 	{
 		auto go{ RenderGOBJ("LevelTile") };
 		m_Tiles[i] = go->AddComponent<TileComp>(TileType::Wall, glm::ivec2{ i % LEVELTILE_COLS, i / LEVELTILE_COLS });
